feat: added 'includeTodoPage' template option
The 'includeTodoPage' template option is a boolean to determine if a "To Do" list page should be generated along with the documentation.
This list will include all `@todo` tags defined anywhere in the system and are categorized by the file in which they appear. A link to the
"to do" list is included in the right navigation